updating build.gradle.kts because danubetech is down. bumping web5-kt again
