+++ httpclient/build.gradle.kts
@@ -7,22 +7,14 @@ plugins {
 
 repositories {
   mavenCentral()
-  // temp maven repo for danubetech
+  // block's cache artifactory for tbd's oss third party dependencies
+  // that do not live in maven central
   maven {
-    name = "tbd-danubetech-temp"
-    url = uri("https://blockxyz.jfrog.io/artifactory/danubetech-temp/")
+    name = "tbd-oss-thirdparty"
+    url = uri("https://blockxyz.jfrog.io/artifactory/tbd-oss-thirdparty-maven2/")
     mavenContent {
       releasesOnly()
     }
-  }
-  maven {
-    url = uri("https://repo.danubetech.com/repository/maven-public")
-  }
-  maven {
-    url = uri("https://jitpack.io")
-  }
-  maven {
-    url = uri("https://repository.jboss.org/nexus/content/repositories/thirdparty-releases/")
   }
 }
